# docs/swagger.yaml
definitions:
...
  internal_handlers.CreateSiteRequest:
    properties:
      name:
        type: string
      org_id:
        type: integer
    required:
    - name
    - org_id
    type: object
...
  internal_handlers.Site:
    properties:
      created_at:
        example: "2024-01-24T10:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Factory 1
        type: string
      org_id:
        example: 1
        type: integer
    type: object
...
  internal_handlers.UpdateSiteRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /api/sites:
    get:
      consumes:
      - application/json
      description: Get a list of sites for the authorized organization
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.Site'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List sites
      tags:
      - sites
    post:
      consumes:
      - application/json
      description: Create a new site for the specified organization
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Site creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateSiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.Site'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new site
      tags:
      - sites
  /api/sites/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a site by ID (cascades to areas, gateways, tags)
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Site ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Site deleted
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Site not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a site
      tags:
      - sites
    get:
      consumes:
      - application/json
      description: Get a single site by ID
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Site ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.Site'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Site not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a site
      tags:
      - sites
    put:
      consumes:
      - application/json
      description: Update a site by ID
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Site ID
        in: path
        name: id
        required: true
        type: integer
      - description: Site update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateSiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.Site'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Site not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a site
      tags:
      - sites
...
swagger: "2.0"